# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "JWT Login API",
        "contact": {
            "name": "Junio Cesar Ferreira"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/login": {
            "post": {
                "description": "Verifica se usuário é registrado e realiza login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Tenta validar usuário e fazer login",
                "parameters": [
                    {
                        "description": "Usuário e Hash",
                        "name": "LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/login.LoginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/protected": {
            "get": {
                "description": "Testando Token JWT",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Rota protegida com autenticação JWT",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "login.LoginData": {
            "type": "object",
            "properties": {
                "pass": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}